refactor: move markdown utility to services directory
- Move src/utils/markdown.js to src/services/markdown-service.js
- Update all import paths in components to use new location
- Remove empty utils directory
- Maintain consistent naming convention with other services
- Fix import paths in Hero2Column, SectionCardCarousel, SectionCards3Column, and SectionImageContent2Column
