+++ src/blocks/Hero2Column/Hero2Column.vue
@@ -70,7 +70,7 @@
 <script setup lang="ts">
   import { computed } from 'vue'
   import Button from '../../components/Button/Button.vue'
-  import { parseMarkdown } from '../../utils/markdown'
+  import { parseMarkdown } from '../../services/markdown-service'
 
   export interface Hero2ColumnButtonProps {
     label: string
